+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - changed value names in summary report
 - introduced dry run mode which overwrites the enforcing settings of each rule. If dry run ist set to true no rule will be enforced.
+- use resource defaults for packages and servies
 
 ## Release 0.9.1